alloy_chains

Struct Chain

pub struct Chain(/* private fields */);

Either a known NamedChain or a EIP-155 chain ID.

Implementations

impl Chain

pub const fn from_named(named: NamedChain) -> Self

Creates a new Chain by wrapping a NamedChain.

pub fn from_id(id: u64) -> Self

Creates a new Chain by wrapping a NamedChain.

pub const fn is_named(self) -> bool

Returns true if this a named variant.

pub const fn is_id(self) -> bool

Returns true if this an Id variant.

pub const fn from_id_unchecked(id: u64) -> Self

Creates a new Chain from the given ID, without checking if an associated NamedChain exists.

This is discouraged, as other methods assume that the chain ID is not known, but it is not unsafe.

pub const fn mainnet() -> Self

Returns the mainnet chain.

pub const fn goerli() -> Self

Returns the goerli chain.

pub const fn holesky() -> Self

Returns the holesky chain.

pub const fn hoodi() -> Self

Returns the hoodi chain.

pub const fn sepolia() -> Self

Returns the sepolia chain.

pub const fn optimism_mainnet() -> Self

Returns the optimism mainnet chain.

pub const fn optimism_goerli() -> Self

Returns the optimism goerli chain.

pub const fn optimism_sepolia() -> Self

Returns the optimism sepolia chain.

pub const fn base_mainnet() -> Self

Returns the base mainnet chain.

pub const fn base_goerli() -> Self

Returns the base goerli chain.

pub const fn base_sepolia() -> Self

Returns the base sepolia chain.

pub const fn arbitrum_mainnet() -> Self

Returns the arbitrum mainnet chain.

pub const fn arbitrum_nova() -> Self

Returns the arbitrum nova chain.

pub const fn arbitrum_goerli() -> Self

Returns the arbitrum goerli chain.

pub const fn arbitrum_sepolia() -> Self

Returns the arbitrum sepolia chain.

pub const fn arbitrum_testnet() -> Self

Returns the arbitrum testnet chain.

pub const fn syndr() -> Self

Returns the syndr l3 mainnet chain.

pub const fn syndr_sepolia() -> Self

Returns the syndr sepolia chain.

pub const fn fraxtal() -> Self

Returns the fraxtal mainnet chain.

pub const fn fraxtal_testnet() -> Self

Returns the fraxtal testnet chain.

pub const fn blast() -> Self

Returns the blast chain.

pub const fn blast_sepolia() -> Self

Returns the blast sepolia chain.

pub const fn linea() -> Self

Returns the linea mainnet chain.

pub const fn linea_goerli() -> Self

Returns the linea goerli chain.

pub const fn linea_sepolia() -> Self

Returns the linea sepolia chain.

pub const fn mode() -> Self

Returns the mode mainnet chain.

pub const fn mode_sepolia() -> Self

Returns the mode sepolia chain.

pub const fn elastos() -> Self

Returns the elastos mainnet chain.

pub const fn degen() -> Self

Returns the degen l3 mainnet chain.

pub const fn dev() -> Self

Returns the dev chain.

pub const fn bsc_mainnet() -> Self

Returns the bsc mainnet chain.

pub const fn bsc_testnet() -> Self

Returns the bsc testnet chain.

pub const fn opbnb_mainnet() -> Self

Returns the opbnb mainnet chain.

pub const fn opbnb_testnet() -> Self

Returns the opbnb testnet chain.

pub const fn ronin() -> Self

Returns the ronin mainnet chain.

pub const fn ronin_testnet() -> Self

Returns the ronin testnet chain.

pub const fn taiko() -> Self

Returns the taiko mainnet chain.

pub const fn taiko_hekla() -> Self

Returns the taiko hekla chain.

pub const fn shimmer() -> Self

Returns the shimmer testnet chain.

pub const fn flare() -> Self

Returns the flare mainnet chain.

pub const fn flare_coston2() -> Self

Returns the flare testnet chain.

pub const fn darwinia() -> Self

Returns the darwinia mainnet chain.

pub const fn crab() -> Self

Returns the crab mainnet chain.

pub const fn koi() -> Self

Returns the koi testnet chain.

pub const fn immutable() -> Self

Returns the Immutable zkEVM mainnet chain.

pub const fn immutable_testnet() -> Self

Returns the Immutable zkEVM testnet chain.

pub const fn ink_sepolia() -> Self

Returns the ink sepolia chain.

pub const fn ink_mainnet() -> Self

Returns the ink mainnet chain.

pub const fn scroll_mainnet() -> Self

Returns the scroll mainnet chain

pub const fn scroll_sepolia() -> Self

Returns the scroll sepolia chain

pub const fn treasure() -> Self

Returns the Treasure mainnet chain.

pub const fn treasure_topaz_testnet() -> Self

Returns the Treasure Topaz testnet chain.

pub const fn berachain() -> Self

Returns the Berachain mainnet chain.

pub const fn berachain_bepolia() -> Self

Returns the Berachain Bepolia testnet chain.

pub const fn sonic() -> Self

Returns the Sonic mainnet chain.

pub const fn sonic_blaze() -> Self

Returns the Sonic Blaze testnet chain.

pub const fn superposition_testnet() -> Self

Returns the Superposition testnet chain.

pub const fn superposition() -> Self

Returns the Superposition mainnet chain.

pub const fn unichain_mainnet() -> Self

Returns the Unichain mainnet chain.

pub const fn unichain_sepolia() -> Self

Returns the Unichain sepolia chain.

pub const fn zksync() -> Self

Returns the ZKSync mainnet chain.

pub const fn zksync_testnet() -> Self

Returns the ZKSync testnet chain.

pub const fn abs() -> Self

Returns the Abstract mainnet chain.

pub const fn abstract_testnet() -> Self

Returns the Abstract testnet chain.

pub const fn sophon() -> Self

Returns the Sophon mainnet chain.

pub const fn sophon_testnet() -> Self

Returns the Sophon testnet chain.

pub const fn lens() -> Self

Returns the Lens mainnet chain.

pub const fn lens_testnet() -> Self

Returns the Lens testnet chain.

pub const fn kind(&self) -> &ChainKind

Returns the kind of this chain.

pub const fn into_kind(self) -> ChainKind

Returns the kind of this chain.

pub const fn is_ethereum(&self) -> bool

Returns true if this chain is Ethereum or an Ethereum testnet.

pub const fn is_optimism(self) -> bool

Returns true if the chain contains Optimism configuration.

pub const fn is_arbitrum(self) -> bool

Returns true if the chain contains Arbitrum configuration.

pub const fn is_elastic(self) -> bool

Returns true if the chain contains Elastic Network configuration.

pub const fn named(self) -> Option<NamedChain>

Attempts to convert the chain into a named chain.

pub const fn id(self) -> u64

The ID of the chain.

impl Chain

Methods delegated to NamedChain. Note that ChainKind::Id won’t be converted because it was already done at construction.

pub const fn average_blocktime_hint(self) -> Option<Duration>

Returns the chain’s average blocktime, if applicable.

See NamedChain::average_blocktime_hint for more info.

pub const fn is_legacy(self) -> bool

Returns whether the chain implements EIP-1559 (with the type 2 EIP-2718 transaction type).

See NamedChain::is_legacy for more info.

pub const fn supports_shanghai(self) -> bool

Returns whether the chain supports the Shanghai hardfork.

See NamedChain::supports_shanghai for more info.

pub const fn etherscan_urls(self) -> Option<(&'static str, &'static str)>

Returns the chain’s blockchain explorer and its API (Etherscan and Etherscan-like) URLs.

See NamedChain::etherscan_urls for more info.

pub const fn etherscan_api_key_name(self) -> Option<&'static str>

Returns the chain’s blockchain explorer’s API key environment variable’s default name.

See NamedChain::etherscan_api_key_name for more info.

pub fn public_dns_network_protocol(self) -> Option<String>

Returns the address of the public DNS node list for the given chain.

See NamedChain::public_dns_network_protocol for more info.