Enum alloy_chains::NamedChain

#[non_exhaustive]

#[repr(u64)]
pub enum NamedChain {

    Mainnet = 1,
    Morden = 2,
    Ropsten = 3,
    Rinkeby = 4,
    Goerli = 5,
    Kovan = 42,
    Holesky = 17_000,
    Sepolia = 11_155_111,
    Alphanet = 41_144_114,
    Optimism = 10,
    OptimismKovan = 69,
    OptimismGoerli = 420,
    OptimismSepolia = 11_155_420,
    Arbitrum = 42_161,
    ArbitrumTestnet = 421_611,
    ArbitrumGoerli = 421_613,
    ArbitrumSepolia = 421_614,
    ArbitrumNova = 42_170,
    Cronos = 25,
    CronosTestnet = 338,
    Rsk = 30,
    Crab = 44,
    Darwinia = 46,
    Koi = 701,
    BinanceSmartChain = 56,
    BinanceSmartChainTestnet = 97,
    Poa = 99,
    Sokol = 77,
    Scroll = 534_352,
    ScrollSepolia = 534_351,
    Metis = 1_088,
    CfxTestnet = 71,
    Cfx = 1_030,
    Gnosis = 100,
    Polygon = 137,
    PolygonMumbai = 80_001,
    PolygonAmoy = 80_002,
    PolygonZkEvm = 1_101,
    PolygonZkEvmTestnet = 1_442,
    Fantom = 250,
    FantomTestnet = 4_002,
    Moonbeam = 1_284,
    MoonbeamDev = 1_281,
    Moonriver = 1_285,
    Moonbase = 1_287,
    Dev = 1_337,
    AnvilHardhat = 31_337,
    GravityAlphaMainnet = 1_625,
    GravityAlphaTestnetSepolia = 13_505,
    Evmos = 9_001,
    EvmosTestnet = 9_000,
    Chiado = 10_200,
    Oasis = 26_863,
    Emerald = 42_262,
    EmeraldTestnet = 42_261,
    FilecoinMainnet = 314,
    FilecoinCalibrationTestnet = 314_159,
    Avalanche = 43_114,
    AvalancheFuji = 43_113,
    Celo = 42_220,
    CeloAlfajores = 44_787,
    CeloBaklava = 62_320,
    Aurora = 1_313_161_554,
    AuroraTestnet = 1_313_161_555,
    Canto = 7_700,
    CantoTestnet = 740,
    Boba = 288,
    Base = 8_453,
    BaseGoerli = 84_531,
    BaseSepolia = 84_532,
    Syndr = 404,
    SyndrSepolia = 444_444,
    Shimmer = 148,
    Fraxtal = 252,
    FraxtalTestnet = 2_522,
    Blast = 81_457,
    BlastSepolia = 168_587_773,
    Linea = 59_144,
    LineaGoerli = 59_140,
    ZkSync = 324,
    ZkSyncTestnet = 300,
    Mantle = 5_000,
    MantleTestnet = 5_001,
    MantleSepolia = 5_003,
    Xai = 660_279,
    XaiSepolia = 37_714_555_429,
    Viction = 88,
    Zora = 7_777_777,
    ZoraGoerli = 999,
    ZoraSepolia = 999_999_999,
    Pgn = 424,
    PgnSepolia = 58_008,
    Mode = 34_443,
    ModeSepolia = 919,
    Elastos = 20,
    KakarotSepolia = 1_802_203_764,
    Etherlink = 42_793,
    EtherlinkTestnet = 128_123,
    Degen = 666_666_666,
    OpBNBMainnet = 204,
    OpBNBTestnet = 5_611,
    Ronin = 2_020,
    Taiko = 167_000,
    TaikoHekla = 167_009,
    AutonomysNovaTestnet = 490_000,
    Flare = 14,
    FlareCoston2 = 114,
    Acala = 787,
    AcalaMandalaTestnet = 595,
    AcalaTestnet = 597,
    Karura = 686,
    KaruraTestnet = 596,
    Pulsechain = 369,
    PulsechainTestnet = 943,
    Immutable = 13_371,
    ImmutableTestnet = 13_473,

}

An Ethereum EIP-155 chain.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Mainnet = 1

Morden = 2

Ropsten = 3

Rinkeby = 4

Goerli = 5

Kovan = 42

Holesky = 17_000

Sepolia = 11_155_111

Alphanet = 41_144_114

Optimism = 10

OptimismKovan = 69

OptimismGoerli = 420

OptimismSepolia = 11_155_420

Arbitrum = 42_161

ArbitrumTestnet = 421_611

ArbitrumGoerli = 421_613

ArbitrumSepolia = 421_614

ArbitrumNova = 42_170

Cronos = 25

CronosTestnet = 338

Rsk = 30

Crab = 44

Darwinia = 46

Koi = 701

BinanceSmartChain = 56

Note the correct name for BSC should be BNB Smart Chain due to the rebranding: https://www.bnbchain.org/en/blog/bsc-is-now-bnb-chain-the-infrastructure-for-the-metafi-universe We keep Binance Smart Chain for backward compatibility, and the enum could be renamed in the future release.

BinanceSmartChainTestnet = 97

Poa = 99

Sokol = 77

Scroll = 534_352

ScrollSepolia = 534_351

Metis = 1_088

CfxTestnet = 71

Cfx = 1_030

Gnosis = 100

Polygon = 137

PolygonMumbai = 80_001

PolygonAmoy = 80_002

PolygonZkEvm = 1_101

PolygonZkEvmTestnet = 1_442

Fantom = 250

FantomTestnet = 4_002

Moonbeam = 1_284

MoonbeamDev = 1_281

Moonriver = 1_285

Moonbase = 1_287

Dev = 1_337

AnvilHardhat = 31_337

GravityAlphaMainnet = 1_625

GravityAlphaTestnetSepolia = 13_505

Evmos = 9_001

EvmosTestnet = 9_000

Chiado = 10_200

Oasis = 26_863

Emerald = 42_262

EmeraldTestnet = 42_261

FilecoinMainnet = 314

FilecoinCalibrationTestnet = 314_159

Avalanche = 43_114

AvalancheFuji = 43_113

Celo = 42_220

CeloAlfajores = 44_787

CeloBaklava = 62_320

Aurora = 1_313_161_554

AuroraTestnet = 1_313_161_555

Canto = 7_700

CantoTestnet = 740

Boba = 288

Base = 8_453

BaseGoerli = 84_531

BaseSepolia = 84_532

Syndr = 404

SyndrSepolia = 444_444

Shimmer = 148

Fraxtal = 252

FraxtalTestnet = 2_522

Blast = 81_457

BlastSepolia = 168_587_773

Linea = 59_144

LineaGoerli = 59_140

ZkSync = 324

ZkSyncTestnet = 300

Mantle = 5_000

MantleTestnet = 5_001

MantleSepolia = 5_003

Xai = 660_279

XaiSepolia = 37_714_555_429

Viction = 88

Zora = 7_777_777

ZoraGoerli = 999

ZoraSepolia = 999_999_999

Pgn = 424

PgnSepolia = 58_008

Mode = 34_443

ModeSepolia = 919

Elastos = 20

KakarotSepolia = 1_802_203_764

Etherlink = 42_793

EtherlinkTestnet = 128_123

Degen = 666_666_666

OpBNBMainnet = 204

OpBNBTestnet = 5_611

Ronin = 2_020

Taiko = 167_000

TaikoHekla = 167_009

AutonomysNovaTestnet = 490_000

Flare = 14

FlareCoston2 = 114

Acala = 787

AcalaMandalaTestnet = 595

AcalaTestnet = 597

Karura = 686

KaruraTestnet = 596

Pulsechain = 369

PulsechainTestnet = 943

Immutable = 13_371

ImmutableTestnet = 13_473

Implementations

impl NamedChain

pub fn as_str(&self) -> &'static str

Returns the string representation of the chain.

pub const fn is_ethereum(&self) -> bool

Returns true if this chain is Ethereum or an Ethereum testnet.

pub const fn is_optimism(self) -> bool

Returns true if the chain contains Optimism configuration.

pub const fn average_blocktime_hint(self) -> Option<Duration>

Returns the chain’s average blocktime, if applicable.

It can be beneficial to know the average blocktime to adjust the polling of an HTTP provider for example.

Note: this is not an accurate average, but is rather a sensible default derived from blocktime charts such as Etherscan’s or Polygonscan’s.

Examples

use alloy_chains::NamedChain;
use std::time::Duration;

assert_eq!(NamedChain::Mainnet.average_blocktime_hint(), Some(Duration::from_millis(12_000)),);
assert_eq!(NamedChain::Optimism.average_blocktime_hint(), Some(Duration::from_millis(2_000)),);

pub const fn is_legacy(self) -> bool

Returns whether the chain implements EIP-1559 (with the type 2 EIP-2718 transaction type).

Examples

use alloy_chains::NamedChain;

assert!(!NamedChain::Mainnet.is_legacy());
assert!(NamedChain::Celo.is_legacy());

pub const fn supports_shanghai(self) -> bool

Returns whether the chain supports the Shanghai hardfork.

pub const fn is_testnet(self) -> bool

Returns whether the chain is a testnet.

pub const fn native_currency_symbol(self) -> Option<&'static str>

Returns the symbol of the chain’s native currency.

pub const fn etherscan_urls(self) -> Option<(&'static str, &'static str)>

Returns the chain’s blockchain explorer and its API (Etherscan and Etherscan-like) URLs.

Returns (API_URL, BASE_URL).

All URLs have no trailing /

Examples

use alloy_chains::NamedChain;

assert_eq!(
    NamedChain::Mainnet.etherscan_urls(),
    Some(("https://api.etherscan.io/api", "https://etherscan.io"))
);
assert_eq!(
    NamedChain::Avalanche.etherscan_urls(),
    Some(("https://api.snowtrace.io/api", "https://snowtrace.io"))
);
assert_eq!(NamedChain::AnvilHardhat.etherscan_urls(), None);

pub const fn etherscan_api_key_name(self) -> Option<&'static str>

Returns the chain’s blockchain explorer’s API key environment variable’s default name.

Examples

use alloy_chains::NamedChain;

assert_eq!(NamedChain::Mainnet.etherscan_api_key_name(), Some("ETHERSCAN_API_KEY"));
assert_eq!(NamedChain::AnvilHardhat.etherscan_api_key_name(), None);

pub fn public_dns_network_protocol(self) -> Option<String>

Returns the address of the public DNS node list for the given chain.

See also https://github.com/ethereum/discv4-dns-lists.