Struct alloy_chains::Chain

pub struct Chain(/* private fields */);

Either a known NamedChain or a EIP-155 chain ID.

Implementations

impl Chain

pub const fn from_named(named: NamedChain) -> Self

Creates a new Chain by wrapping a NamedChain.

pub fn from_id(id: u64) -> Self

Creates a new Chain by wrapping a NamedChain.

pub const fn from_id_unchecked(id: u64) -> Self

Creates a new Chain from the given ID, without checking if an associated NamedChain exists.

This is discouraged, as other methods assume that the chain ID is not known, but it is not unsafe.

pub const fn mainnet() -> Self

Returns the mainnet chain.

pub const fn goerli() -> Self

Returns the goerli chain.

pub const fn holesky() -> Self

Returns the holesky chain.

pub const fn sepolia() -> Self

Returns the sepolia chain.

pub const fn optimism_mainnet() -> Self

Returns the optimism mainnet chain.

pub const fn optimism_goerli() -> Self

Returns the optimism goerli chain.

pub const fn optimism_sepolia() -> Self

Returns the optimism sepolia chain.

pub const fn base_mainnet() -> Self

Returns the base mainnet chain.

pub const fn base_goerli() -> Self

Returns the base goerli chain.

pub const fn base_sepolia() -> Self

Returns the base sepolia chain.

pub const fn syndr() -> Self

Returns the syndr l3 mainnet chain.

pub const fn syndr_sepolia() -> Self

Returns the syndr sepolia chain.

pub const fn fraxtal() -> Self

Returns the fraxtal mainnet chain.

pub const fn fraxtal_testnet() -> Self

Returns the fraxtal testnet chain.

pub const fn blast() -> Self

Returns the blast chain.

pub const fn blast_sepolia() -> Self

Returns the blast sepolia chain.

pub const fn linea() -> Self

Returns the linea mainnet chain.

pub const fn linea_goerli() -> Self

Returns the linea goerli chain.

pub const fn mode() -> Self

Returns the mode mainnet chain.

pub const fn mode_sepolia() -> Self

Returns the mode sepolia chain.

pub const fn elastos() -> Self

Returns the elastos mainnet chain.

pub const fn degen() -> Self

Returns the degen l3 mainnet chain.

pub const fn dev() -> Self

Returns the dev chain.

pub const fn bsc_mainnet() -> Self

Returns the bsc mainnet chain.

pub const fn bsc_testnet() -> Self

Returns the bsc testnet chain.

pub const fn opbnb_mainnet() -> Self

Returns the opbnb mainnet chain.

pub const fn opbnb_testnet() -> Self

Returns the opbnb testnet chain.

pub const fn ronin() -> Self

Returns the ronin mainnet chain.

pub const fn taiko() -> Self

Returns the taiko mainnet chain.

pub const fn taiko_hekla() -> Self

Returns the taiko hekla chain.

pub const fn shimmer() -> Self

Returns the shimmer testnet chain.

pub const fn flare() -> Self

Returns the flare mainnet chain.

pub const fn flare_coston2() -> Self

Returns the flare testnet chain.

pub const fn darwinia() -> Self

Returns the darwinia mainnet chain.

pub const fn crab() -> Self

Returns the crab mainnet chain.

pub const fn koi() -> Self

Returns the koi testnet chain.

pub const fn immutable() -> Self

Returns the Immutable zkEVM mainnet chain.

pub const fn immutable_testnet() -> Self

Returns the Immutable zkEVM testnet chain.

pub const fn kind(&self) -> &ChainKind

Returns the kind of this chain.

pub const fn into_kind(self) -> ChainKind

Returns the kind of this chain.

pub const fn is_ethereum(&self) -> bool

Returns true if this chain is Ethereum or an Ethereum testnet.

pub const fn is_optimism(self) -> bool

Returns true if the chain contains Optimism configuration.

pub const fn named(self) -> Option<NamedChain>

Attempts to convert the chain into a named chain.

pub const fn id(self) -> u64

The ID of the chain.

impl Chain

Methods delegated to NamedChain. Note that ChainKind::Id won’t be converted because it was already done at construction.

pub const fn average_blocktime_hint(self) -> Option<Duration>

Returns the chain’s average blocktime, if applicable.

See NamedChain::average_blocktime_hint for more info.

pub const fn is_legacy(self) -> bool

Returns whether the chain implements EIP-1559 (with the type 2 EIP-2718 transaction type).

See NamedChain::is_legacy for more info.

pub const fn supports_shanghai(self) -> bool

Returns whether the chain supports the Shanghai hardfork.

See NamedChain::supports_shanghai for more info.

pub const fn etherscan_urls(self) -> Option<(&'static str, &'static str)>

Returns the chain’s blockchain explorer and its API (Etherscan and Etherscan-like) URLs.

See NamedChain::etherscan_urls for more info.

pub const fn etherscan_api_key_name(self) -> Option<&'static str>

Returns the chain’s blockchain explorer’s API key environment variable’s default name.

See NamedChain::etherscan_api_key_name for more info.

pub fn public_dns_network_protocol(self) -> Option<String>

Returns the address of the public DNS node list for the given chain.

See NamedChain::public_dns_network_protocol for more info.